% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{add_var}
\alias{add_var}
\title{Add variable if it doesn't exist}
\usage{
add_var(df, x)
}
\arguments{
\item{df}{dataframe}

\item{x}{character of new variable name}
}
\value{
data frame with new variable
}
\description{
Add variable if it doesn't exist
}
\keyword{internal}
