% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_partner_types.R
\name{apply_partner_type}
\alias{apply_partner_type}
\title{Apply partner type (local/international)}
\usage{
apply_partner_type(df)
}
\arguments{
\item{df}{data frame, MER or financial dataset}
}
\value{
a df with the partner types provided by USAID.
}
\description{
`apply_partner_type` pulls from USAID managed Google Sheet maintained by the
Local Partner team that designates all USAID mechanisms by their type, local
or international. The designations include the updates for regional
partners, which in FY21 are classified as local. These adjustments can be
found on the adjusted column.
}
