% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{browse_knowissues}
\alias{browse_knowissues}
\title{Browse Known Issues}
\usage{
browse_knowissues()
}
\value{

}
\description{
Launch Known Data Issues Tracker
}
\examples{
 \dontrun{
 browse_knowissues() }
}
