% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{browse_knownissues}
\alias{browse_knownissues}
\title{Browse Known Issues}
\usage{
browse_knownissues()
}
\value{

}
\description{
Launch Known Data Issues Tracker
}
\examples{
 \dontrun{
 browse_knownissues() }
}
