% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_agency}
\alias{clean_agency}
\title{Clean data from fundingagency column}
\usage{
clean_agency(.data)
}
\arguments{
\item{.data}{MSD Datasets}
}
\value{
Cleaned MSD DataFrame
}
\description{
Clean data from fundingagency column
}
\examples{
\dontrun{
 df_msd \%>\% clean_agency() }
}
