% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_countries}
\alias{clean_countries}
\title{Clean OU/Country names to match PEPFAR Data}
\usage{
clean_countries(.data, colname = "admin")
}
\arguments{
\item{.data}{MSD Datasets}

\item{colname}{Column name to be updated}
}
\value{
Cleaned DataFrame
}
\description{
Clean OU/Country names to match PEPFAR Data
}
\examples{
\dontrun{
 spdf <- ne_countries(type = "sovereignty", scale = 110, returnclass = "sf") \%>\%
   dplyr::select(sovereignt, admin, name, adm0_a3) \%>\%
   glamr::clean_countries(colname = "admin") }
}
