% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_filename}
\alias{clean_filename}
\title{Clean Filename}
\usage{
clean_filename(
  x,
  rm_apostrophe = TRUE,
  rp_space = FALSE,
  mk_lower = FALSE,
  add_date = NULL
)
}
\arguments{
\item{x}{filepath or file name}

\item{rm_apostrophe}{remove all apostrophes, default = TRUE}

\item{rp_space}{replace spaces with underscore, default = FALSE}

\item{mk_lower}{make lowercase, default = FALSE}

\item{add_date}{add date "prefix" or "suffix"}
}
\value{
clean filename
}
\description{
This function is primarily useful for removing any apostrophe from the filename
since this will get rejected by Google Drive, but also includes features like
replacing spaces with an underscore, converting to all lowercase, and adding
a date prefix or suffix.
}
