% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_indicator}
\alias{clean_indicator}
\title{Clean indicators (apply _D suffix)}
\usage{
clean_indicator(df)
}
\arguments{
\item{df}{MSD data frame}
}
\value{
indicators with denominator have _D suffix
}
\description{
`clean_indicator` applies a '_D' suffix to any indicators that are a
denominator. This is particularly useful when aggregating data or
reshaping.
}
\examples{
\dontrun{
df <- df \%>\%
filter(indicator == "TX_PVLS",
       standardizeddisaggregate \%in\% c("Total Numerator", "Total Denominator")) \%>\%
clean_indicator() }
}
\seealso{
Other column munging: 
\code{\link{clean_agency}()},
\code{\link{clean_column}()},
\code{\link{clean_countries}()},
\code{\link{clean_psnu}()}
}
\concept{column munging}
