% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{clean_knownissues}
\alias{clean_knownissues}
\title{Clean Known Issues Google Sheet}
\usage{
clean_knownissues(df)
}
\arguments{
\item{df}{data from get_knownissues()}
}
\value{
adjust mech_code and tidy by expanding by indicator
}
\description{
Clean Known Issues Google Sheet
}
