% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{complete_exclude_vars}
\alias{complete_exclude_vars}
\title{Add full set of exclusion variables}
\usage{
complete_exclude_vars(df)
}
\arguments{
\item{df}{df output from flag_knownissues()}
}
\value{
data frame with new variables
}
\description{
Add full set of exclusion variables
}
\keyword{internal}
