% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{connect_text}
\alias{connect_text}
\title{Clean and connect parts of text together}
\usage{
connect_text(txt, connections = "[^a-zA-Z0-9]", connector = "_")
}
\arguments{
\item{txt}{String charactors}

\item{connections}{Characters to be replaced by connector}

\item{connector}{Charactor used as connector}
}
\value{
cleaned text
}
\description{
Clean and connect parts of text together
}
\examples{
\dontrun{
connect_text("THIS - is complex (very bad)")}

}
