% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_date_period.R
\name{convert_date_to_qtr}
\alias{convert_date_to_qtr}
\title{Convert Date to FY Quarter/Period}
\usage{
convert_date_to_qtr(date)
}
\arguments{
\item{date}{date formatted like 2021-10-01}
}
\value{
vector of FY period, eg FY22Q1
}
\description{
Convert Date to FY Quarter/Period
}
\examples{
\dontrun{
dates <- c("2021-10-01", "2021-11-15")
convert_date_to_qtr(dates) }

}
\seealso{
Other period: 
\code{\link{convert_datim_pd_to_qtr}()},
\code{\link{convert_fy_qtr_to_pd}()},
\code{\link{convert_qtr_to_date}()}
}
\concept{period}
