% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_date_period.R
\name{convert_datim_pd_to_qtr}
\alias{convert_datim_pd_to_qtr}
\title{Convert DATIM CY Quarter/Period to FY Quarter}
\usage{
convert_datim_pd_to_qtr(df, pd_col = "Period")
}
\arguments{
\item{df}{dataframe from DATIM API, \code{grabr::extract_datim()}}

\item{pd_col}{name of the period column, default = "Period"}
}
\value{
Convert periods from long CY dates to PEPFAR standard FY
}
\description{
Convert a period a DATIM API in the format of FY22Q1 or FY22 (for
targets/cumulative). This function is built into `extract_datim`.
}
\examples{
\dontrun{
df <- tibble::tibble(Periods = c("October - December 2023",
                                 "January - March 2024",
                                 "October 2023 to September 2024"))
df <- convert_datim_pd_to_qtr(df) }
}
\seealso{
[set_datim()] to store DATIM authentication;
[load_secrets()] to store DATIM authentication

Other period: 
\code{\link{convert_date_to_qtr}()},
\code{\link{convert_fy_qtr_to_pd}()},
\code{\link{convert_qtr_to_date}()}
}
\concept{period}
