% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_date_period.R
\name{convert_qtr_to_date}
\alias{convert_qtr_to_date}
\title{Convert FY Quarter/Period to Date}
\usage{
convert_qtr_to_date(period, type = "start")
}
\arguments{
\item{period}{period formated like FY22Q1 or FY22}

\item{type}{start or end date of quarter/period, default = "start"}
}
\value{
date vector
}
\description{
Convert a period (from reshape_msd()) in the format of FY22Q1 or FY22 (for
targets/cumulative).
}
\examples{
\dontrun{
df <- read_msd(path)
df <- df \%>\%
  filter(df,
         operatingunit == "Jupiter",
         indicator == "TX_NEW",
         standarddisaggregate == "Total Numerator") \%>\%
  group_by(fiscal_year, primepartner) \%>\%
  summarize(across(start_with("qtr"), sum, na.rm = TRUE)) \%>\%
  ungroup()
df <- df \%>\%
  reshape_msd() \%>\%
  mutate(date = convert_qtr_to_date(period), .after = period) }

}
\seealso{
Other period: 
\code{\link{convert_date_to_qtr}()},
\code{\link{convert_datim_pd_to_qtr}()},
\code{\link{convert_fy_qtr_to_pd}()}
}
\concept{period}
