% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{curr_date}
\alias{curr_date}
\title{Get formatted current date}
\usage{
curr_date(fmt = "\%Y-\%m-\%d")
}
\description{
Get formatted current date
}
\examples{
\dontrun{
 curr_date()
 curr_date(fmt = "\%m/\%d/\%Y")
}
}
