% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dim_items}
\alias{datim_dim_items}
\title{Get PEPFAR/DATIM Dimension Items}
\usage{
datim_dim_items(
  dimension,
  var = NULL,
  url = "https://final.datim.org/api/dimensions",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{dimension}{Dimension name}

\item{var}{column name to pull values from, id or item}

\item{url}{DATIM API end point}

\item{username}{DATIM Account Username}

\item{password}{DATIM Account passward}
}
\value{
Dimension's items as tibble or vector
}
\description{
Get PEPFAR/DATIM Dimension Items
}
\examples{
\dontrun{
  library(glamr)

  datim_dim_items(dimension = "Funding Agency")
  datim_dim_items(dimension = "Funding Agency", var = "item")
}

}
