% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dim_url}
\alias{datim_dim_url}
\title{Build PEPFAR/DATIM Query dimension}
\usage{
datim_dim_url(
  dimension,
  items = NULL,
  url = "https://final.datim.org/api/dimensions",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{dimension}{Dimension name}

\item{items}{Item name}

\item{url}{DATIM API End point}

\item{username}{DATIM Account Username}

\item{password}{DATIM Account passward}
}
\value{
Valid DATIM Query Params url
}
\description{
Build PEPFAR/DATIM Query dimension
}
\examples{
\dontrun{
  library(glamr)

  datim_dim_url(dimension = "Sex")

  datim_dim_url(
    dimension = "Disaggregation Type",
    items = "Age/Sex/HIVStatus"
  )

  datim_dim_url(
    dimension = "Disaggregation Type",
    items = c("Age/Sex", "Age/Sex/HIVStatus")
  )

}


}
