% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dimension}
\alias{datim_dimension}
\title{Get PEPFAR/DATIM Dimension ID}
\usage{
datim_dimension(name, username = NULL, password = NULL)
}
\arguments{
\item{name}{Dimension name}

\item{username}{DATIM Account Username}

\item{password}{DATIM Account passward}
}
\value{
dimension uid
}
\description{
Get PEPFAR/DATIM Dimension ID
}
\examples{
\dontrun{
  library(glamr)
  datim_dimension("OU Level")
}

}
