% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_execute_query}
\alias{datim_execute_query}
\title{Execute Datim Query}
\usage{
datim_execute_query(url, username = NULL, password = NULL, flatten = FALSE)
}
\arguments{
\item{url}{API Base url & all query parameters}

\item{username}{Datim username, recommend using `glamr::datim_user()`}

\item{password}{Datim password, recommend using `glamr::datim_pwd()`}

\item{flatten}{Should query json result be flatten? Default is false}
}
\value{
returns query results as json object
}
\description{
Execute Datim Query
}
\examples{
\dontrun{
  library(glamr)

  execute_datim_query(
    url = 'https://www.datim.org/api/sqlViews/<uid>?format=json',
    username = datim_user(),
    password = datim_pwd(),
    flatten = TRUE
  )
}

}
