% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_pops}
\alias{datim_pops}
\title{Extract PLHIV and General POP Estimates from datim}
\usage{
datim_pops(
  ou,
  level = "country",
  fy = NULL,
  hierarchy = FALSE,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{ou}{Operatingunit}

\item{level}{Organization level}

\item{fy}{Fiscal Year}

\item{hierarchy}{Should additional organizational hierarchy be added?, default is FALSE}

\item{username}{Datim account username}

\item{password}{Datim account password}
}
\value{
PLHIV and POP_EST Data
}
\description{
Extract PLHIV and General POP Estimates from datim
}
\examples{
\dontrun{
  library(glamr)

  datim_pops(ou = "Nigeria")

  datim_pops(ou = "Nigeria", fy = 2021)

  datim_pops(ou = "Nigeria", level = "psnu", fy = 2021, hierarchy = TRUE)
}
}
