% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_process_query}
\alias{datim_process_query}
\title{Process Datim Query results}
\usage{
datim_process_query(url, username = NULL, password = NULL)
}
\arguments{
\item{username}{Datim username, recommend using `glamr::datim_user()`}

\item{password}{Datim password, recommend using `glamr::datim_pwd()`}
}
\value{
Data as tibble
}
\description{
Process Datim Query results
}
\examples{
\dontrun{
 library(glamr)
 datim_process_query("<full-api-call-url>")
}

}
