% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{datim_pwd}
\alias{datim_pwd}
\title{Return DATIM password}
\usage{
datim_pwd()
}
\value{
access DATIM password from keyring
}
\description{
Return DATIM password
}
\examples{
\dontrun{
load_secrets()
user <- getOption("datim")
ou_table <- datim_outable(user, datim_pwd()) }
}
