% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_query}
\alias{datim_query}
\title{Query PEPFAR/DATIM targets/results data}
\usage{
datim_query(
  ou,
  level = "prioritization",
  pe = "THIS_FINANCIAL_YEAR",
  ta = "PLHIV",
  value = NULL,
  disaggs = NULL,
  dimensions = NULL,
  property = "SHORTNAME",
  metadata = TRUE,
  hierarchy = TRUE,
  baseurl = "https://final.datim.org/api",
  username = NULL,
  password = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ou}{Operatingunit}

\item{level}{Organization hierarchy level}

\item{pe}{Reporting period. This can be expressed as relative or fixed periods
Eg.: "THIS_FINANCIAL_YEAR", "2020Oct", "QUARTERS_THIS_YEAR", "2021Q2"
default is "THIS_FINANCIAL_YEAR"}

\item{ta}{Technical Area, valid option can be obtain from `datim_dim_items("Technical Area")`}

\item{value}{Type of value to return, MER Targets or Results or both}

\item{disaggs}{Disaggregation Types. This depends on the value of ta}

\item{dimensions}{Additional dimensions and/or columns. This depends on values of ta and disaggs}

\item{property}{Type of name}

\item{metadata}{Should metadata be included}

\item{baseurl}{DATIM API End point url}

\item{username}{Datim username}

\item{password}{Datim username}

\item{verbose}{Display all notifications}

\item{hierachy}{Should additional hirarchy level be included}
}
\value{
data as tibble
}
\description{
Query PEPFAR/DATIM targets/results data
}
\examples{
\dontrun{
 library(glamr)

 datim_query(ou = "Nigeria", ta = "PLHIV")

 datim_query(ou = "Nigeria", ta = "POP_EST")

 datim_query(ou = "Nigeria",
             level = "country",
             ta = "TX_CURR",
             disaggs = "Age/Sex/HIVStatus",
             dimensions = c("Age: <15/15+  (Coarse)", "Sex"))
}

}
