% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_query_old}
\alias{datim_query_old}
\title{Query targets data from DATIM}
\usage{
datim_query_old(
  ou,
  org_lvl = "facility",
  type = "TX",
  datim_cy = Wavelength::curr_fy - 1,
  username,
  password,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ou}{Operating Unit}

\item{org_lvl}{Level}

\item{type}{Technical}

\item{datim_cy}{Fiscal year}

\item{username}{Datim Username}

\item{password}{Datim Password}

\item{baseurl}{Datim API URL}
}
\description{
Query targets data from DATIM
}
\note{
Old function re-written as `datim_query()``
}
