% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{extract_datim}
\alias{extract_datim}
\title{DATIM Analytics API}
\usage{
extract_datim(url, username, password)
}
\arguments{
\item{url}{supply url for API call}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using `mypwd()`}
}
\description{
DATIM Analytics API
}
\examples{
\dontrun{
 myurl <- paste0(baseurl, "api/29/analytics.json?
                 dimension=LxhLO68FcXm:udCop657yzi&
                 dimension=ou:LEVEL-4;HfVjCurKxh2&
                 filter=pe:2018Oct&
                 displayProperty=SHORTNAME&outputIdScheme=CODE")
 myuser <- "UserX"
 df_targets <- get_datim_targets(myurl, myuser, mypwd(myuser)) }
}
