% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_excel_data.R
\name{extract_excel_data}
\alias{extract_excel_data}
\title{Extract data from excel link}
\usage{
extract_excel_data(src_page, link_id, file_sheet = 2, file_ext = "xlsx")
}
\arguments{
\item{src_page}{The http(s)link to the source web page}

\item{link_id}{A CSS identifer of the hyperlinked element}

\item{file_sheet}{The file sheet number or name}

\item{file_ext}{The extension of the file}
}
\value{
File content as a data frame
}
\description{
Extract data from excel link
}
