% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tbl_data.R
\name{extract_tbl_data}
\alias{extract_tbl_data}
\title{Extract table data from web page}
\usage{
extract_tbl_data(src_url, tbl_id)
}
\arguments{
\item{src_url}{The http(s)link to the source web page}

\item{tbl_id}{A unique identifer of the target table}
}
\value{
A data frme
}
\description{
Extract table data from web page
}
