% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_text.R
\name{extract_text}
\alias{extract_text}
\title{Extract text}
\usage{
extract_text(txt, limits = "()")
}
\arguments{
\item{txt}{text containing parenthesis}

\item{limits}{area to extract text from, c("()", "{}", "[]")}
}
\value{
text within limits
}
\description{
Extract text
}
\examples{
\dontrun{
 extract_text(txt = "Saint Mary Hopital (SMH)")
 extract_text(txt = "TDB [Placeholder - New Mechnism]") }
}
