% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{flag_knownissues}
\alias{flag_knownissues}
\title{Flag Known Issues}
\usage{
flag_knownissues(df, df_issues)
}
\arguments{
\item{df}{original MSD input}

\item{df_issues}{df output from clean_knownissues()}
}
\value{
df of MSD with extra column of period type to know what to excluded
}
\description{
Flag Known Issues
}
\keyword{internal}
