% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{folder_setup}
\alias{folder_setup}
\title{Setup Folder Structure in Project}
\usage{
folder_setup(
  folder_list = list("Data", "Images", "Scripts", "AI", "Dataout", "GIS", "Documents",
    "Graphics", "markdown")
)
}
\arguments{
\item{folder_list}{list of folders to install}
}
\description{
`folder_setup()` creates an organizational structure that is common across
OHA/SI projects so every analyst knows what to expect and where when picking
up a new project or one cloned from a co-worker. This function can be used
as a stand alone function but primarily serves `si_setup()`
}
\details{
The standard setup provides the following folders for these uses:
  * Data - where any raw/input data (**xlsx/csv/rds**) specific to the project are stored
  * Dataout - where any intermediary or final data (**xlsx/csv/rds**) are output as a product of your code
  * Scripts - where all the code (**R/py**) are stored (if there is a local order, make sure to add prefixes to each script, e.g. 00_init.R, 01_data-access.R, 02_data-munging.R, ...)
  * Images - any **png/jpeg** visual outputs from your code
  * Graphics - any **svg/pdf** visual outputs that will be edited in vector graphics editor, eg Adobe Illustrator or Inkscape
  * AI - any **ai** files or other files from a graphics editor (exported pngs products will be stored in Images)
  * GIS - any **shp** files or other GIS releated inputs
  * Documents - any **docx/xlsx/pptx/pdf** documents that relate to the process or are final outputs
  * markdown - exported **md** files from a knitr report
}
\examples{
\dontrun{
#standard
  folder_setup()
#specific
  fldrs <- c("Data", "Tableau", "AI")
  folder_setup(fldrs)
}
}
\seealso{
Other project setup: 
\code{\link{setup_gitignore}()},
\code{\link{setup_readme}()},
\code{\link{si_setup}()}
}
\concept{project setup}
