% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_drivefile.R
\name{gdrive_folder}
\alias{gdrive_folder}
\title{Get id of googledrive folder}
\usage{
gdrive_folder(name, path = NULL, add = FALSE, ...)
}
\arguments{
\item{name}{Googledrive folder name}

\item{path}{Googledrive parent path id}

\item{add}{Should folder be added if missing, default is true}

\item{...}{Other arguments to passed on to `drive_mkdir`}
}
\value{
Googledrive folder item it or NULL for non existing folder
}
\description{
Get id of googledrive folder
}
\note{
This function will create a new folder if add is set to TRUE
}
\examples{
\dontrun{
 library(glamr)

 gdrive_folder("Test-Folder", "ID-adfdfsdfdfdfs")
}

}
