% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{get_account}
\alias{get_account}
\title{Get account details}
\usage{
get_account(name)
}
\arguments{
\item{name}{Service name of the account}
}
\value{
key / value pair as list containing details of the account (invisible)
}
\description{
Get account details
}
\note{
Inspired by `grabr::lazy_secrets()`
}
\examples{
\dontrun{
get_account(name = 's3')}

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_bucket}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
