% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{get_keys}
\alias{get_keys}
\title{Get Service Keys}
\usage{
get_keys(service)
}
\arguments{
\item{service}{Account Service name}
}
\value{
list of key names for active services
}
\description{
Get Service Keys
}
\examples{
\dontrun{
get_keys('<service-name>')}

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
