% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_levels}
\alias{get_levels}
\title{Get all orgunits levels in org hierarchy}
\usage{
get_levels(
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{username}{DATIM username, recommed using `datim_user()`}

\item{password}{DATIM password, recommend using `datim_pwd()`}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\value{
df
}
\description{
Get all orgunits levels in org hierarchy
}
\note{
Similar to `glamr::identify_levels()` and `glamr::get_outable()`
}
\examples{
\dontrun{
  library(glamr)

  # Get PEPFAR Org Levels
  get_levels()
 }

}
