% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_ouorglevel}
\alias{get_ouorglevel}
\title{Get OU Org level}
\usage{
get_ouorglevel(
  operatingunit,
  country = NULL,
  org_type = "prioritization",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{operatingunit}{Operatingunit name}

\item{country}{Country name (default = Operatingunit)}

\item{org_type}{Orgunit type (country_lvl, prioritization, community, facility_lvl)}

\item{username}{Datim Account username}

\item{password}{Datim Account Password}
}
\value{
uid
}
\description{
Get OU Org level
}
\examples{
\dontrun{
 library(gisr)

 cntry <- "Zambia"

 # Get country org level
 get_ouorglevel(cntry)

 # Get community org level
 get_ouorglevel(cntry, org_type = "community")
}

}
