% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_ouorgs}
\alias{get_ouorgs}
\title{Get list of OU Orgs at specific level}
\usage{
get_ouorgs(
  ouuid,
  level = 4,
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ouuid}{OU uid}

\item{level}{org level}

\item{username}{DATIM Username}

\item{password}{DATIM password, recommend using `mypwd()`}

\item{baseurl}{base url for the API, default = https://final.datim.org/}
}
\value{
ORG UIDS as tibble
}
\description{
Get list of OU Orgs at specific level
}
\note{
Use `get_orguids()` for levels above 4
}
\examples{
\dontrun{

 library(glamr)

 cntry <- "Zambia"

 uid <- get_ouuid(cntry)

 lvl <- get_ouorglevel(cntry, org_type = "prioritization")

 orgs <- get_ouorgs(ouuid = uid, level = lvl)
}

}
