% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{get_s3key}
\alias{get_s3key}
\title{Get S3 Credentials - Access or Secret Access Key}
\usage{
get_s3key(name = "access")
}
\arguments{
\item{name}{S3 account key}
}
\value{
stored key
}
\description{
`get_s3key` retrieves your S3 keys using the `keyring` package.
Set name to `access` for `Access Key`,
`name` to `secret` for `Secret Access Key`
}
\examples{
\dontrun{
get_s3key(name = "access") }

}
