% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{get_services}
\alias{get_services}
\title{Get Services}
\usage{
get_services()
}
\value{
list of active services
}
\description{
Get Services
}
\examples{
\dontrun{
get_services()}

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_access}()},
\code{\link{pdap_bucket}()},
\code{\link{pdap_secret}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
