% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glamr-package.R
\docType{package}
\name{glamr-package}
\alias{glamr}
\alias{glamr-package}
\title{glamr: SI Utilities Package}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a series of base functions useful to the GH OHA SI team. This includes project setup, pulling from DATIM, and key functions for working with the MSD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/glamr}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/glamr/issues}
}

}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.gov}

Authors:
\itemize{
  \item Tim Essam \email{tessam@usaid.gov}
  \item Baboyma Kagniniwa \email{bkagniniwa@usaid.gov}
  \item Benjamin Kasdan \email{bkasdan@usaid.gov}
}

}
\keyword{internal}
