% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\docType{data}
\name{gs_id_knownissues}
\alias{gs_id_knownissues}
\title{Known Data Issues Tracker Google Sheet ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
gs_id_knownissues
}
\value{
google sheet id
}
\description{
Known Data Issues Tracker Google Sheet ID
}
\keyword{datasets}
