% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{identify_levels}
\alias{identify_levels}
\title{Identify Facility/Community levels in org hierarchy}
\usage{
identify_levels(username, password, baseurl = "https://final.datim.org/")
}
\arguments{
\item{username}{DATIM username}

\item{password}{DATIM password, recommend using `mypwd()`}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\description{
Identify Facility/Community levels in org hierarchy
}
\examples{
\dontrun{
 #table for all OUs
  myuser <- "UserX"
  identify_levels(username = myuser, password = mypwd(myuser)) }
}
