% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{identify_pd}
\alias{identify_pd}
\title{Identify Reporting Period from MSD File}
\usage{
identify_pd(msd_file, clean = TRUE)
}
\arguments{
\item{msd_file}{MSD File, should include publication date}

\item{clean}{Should be period be clean? default is true}
}
\value{
reporting period
}
\description{
Identify Reporting Period from MSD File
}
\examples{
\dontrun{

 library(glamr)

 pd = return_latest(si_path(), "OU_IM") \%>\% identify_pd()
}

}
