% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_installed}
\alias{is_installed}
\title{Test if package is installed locally}
\usage{
is_installed(package)
}
\arguments{
\item{package}{name of R package to check}
}
\value{
A boolean
}
\description{
Test if package is installed locally
}
