% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_loaded}
\alias{is_loaded}
\title{Test if option variable is loaded in current session}
\usage{
is_loaded(opt_var)
}
\arguments{
\item{opt_var}{option variable ("datim", "email", "path_msd", "path_datim", "path_vector", "path_raster")}
}
\value{
A boolean
}
\description{
Test if option variable is loaded in current session
}
