% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_pdap}
\alias{is_pdap}
\title{Test if the user is working on the PDAP environment in Posit Workbench}
\usage{
is_pdap()
}
\value{
a boolean
}
\description{
Test if the user is working on the PDAP environment in Posit Workbench
}
\keyword{internal}
