% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{load_secrets}
\alias{load_secrets}
\title{Load credentials}
\usage{
load_secrets(service = c("email", "datim", "pano", "s3", "pdap"))
}
\arguments{
\item{service}{account, either "email", "datim", "pano", "s3", or "pdap"; by
default, all are loaded if they are available}
}
\value{
stores Google, DATIM, PEFPFAR Panorama, s3, and PDAP credentials in
  session
}
\description{
`load_secrets` should be set at the beginning of a script to store your
 email and DATIM user name under Options for the current session. This allows
 analysts to more easily share their scripts without having to manually update
 or remove use names.
}
\details{
To initially store your credentials, you will first need to
 run `set_email()`, `set_datim()`, `set_pano()`, and/or `set_key` (for s3)

 `load_secrets` utilizes `keyring` package to access the OS credentials store.
 Storing in a centralized, secure location allows analysts to other analysts code
 without having to manually change user names/email address to access
 DATIM or Google Drive.
}
\examples{
\dontrun{
load_secrets()
ou_table <- datim_outable(datim_user(), datim_pwd()) }
}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_access}()},
\code{\link{pdap_bucket}()},
\code{\link{pdap_secret}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
