% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{lookup_country}
\alias{lookup_country}
\title{Lookup Official Country name}
\usage{
lookup_country(country, language = "en")
}
\arguments{
\item{country}{country name}

\item{language}{language to use for lookup}
}
\value{
cleaned country name
}
\description{
Lookup Official Country name
}
\keyword{internal}
