% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{missing_section}
\alias{missing_section}
\title{Check if standard text section is missing from file}
\usage{
missing_section(file)
}
\arguments{
\item{file}{either README.md or .gitignore}
}
\description{
Check if standard text section is missing from file
}
\keyword{internal}
