% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{note_knownissues}
\alias{note_knownissues}
\title{Note Known Issues}
\usage{
note_knownissues(df_orig, df_allissues)
}
\arguments{
\item{df_orig}{MSD dataset}

\item{df_allissues}{df from clean_knownissues()}
}
\value{
df that has a description of what was excluded in orig df
}
\description{
Note Known Issues
}
