% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_content}
\alias{pano_content}
\title{Read html content from web page}
\usage{
pano_content(page_url, session)
}
\arguments{
\item{page_url}{URL of the page to be extracted}

\item{session}{Valid and active login session}
}
\value{
html content
}
\description{
Read html content from web page
}
\examples{
\dontrun{
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
  url <- "https://pepfar-panorama.org/forms/downloads"

  cont <- pano_content(page_url = url, session = s)
}

}
