% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_download}
\alias{pano_download}
\title{Download file from PEPFAR Panorama}
\usage{
pano_download(item_url, session, dest = NULL, uncompress = FALSE)
}
\arguments{
\item{item_url}{URL for the item to be downlaoded}

\item{session}{Login session}

\item{dest}{Location and name of the destination file}

\item{uncompress}{If yes, the downloaded zip file will be decompressed. Default is FALSE}
}
\value{
file content as binary
}
\description{
Download file from PEPFAR Panorama
}
