% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_elements}
\alias{pano_elements}
\title{Extract data elements details from html content}
\usage{
pano_elements(page_html, page_name = NA, page_url = NULL)
}
\arguments{
\item{page_html}{html content}

\item{page_name}{Current html page name}

\item{page_url}{Curretn html page url}
}
\value{
html element
}
\description{
Extract data elements details from html content
}
\examples{
\dontrun{
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
  url <- "https://pepfar-panorama.org/forms/downloads"

  cont <- pano_content(page_url = url, session = s)

  elts <- pano_elements(page_html = cont, page_url = url)
}

}
