% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_extract}
\alias{pano_extract}
\title{Extract data outputs from Panorama}
\usage{
pano_extract(
  item = "mer",
  version = "initial",
  fiscal_year = 2021,
  quarter = 3,
  unpack = FALSE,
  username = NULL,
  password = NULL,
  base_url = NULL
)
}
\arguments{
\item{item}{Panorama data type. Eg: mer, financial, sims, narratives}

\item{version}{Data release version: initial or clean}

\item{fiscal_year}{Reporting Fiscal year}

\item{quarter}{Reporting Quarter}

\item{unpack}{If yes, unpack nested directories}

\item{username}{Panorama username, recommend using `pano_user()`}

\item{password}{Panorama password, recommend using `pano_pwd()`}

\item{base_url}{Panorama base url}
}
\value{
list of output files as data frame
}
\description{
Extract data outputs from Panorama
}
\note{
This function combines `pano_session()`, `pano_content()`, `pano_elements()`, and in some cases `pano_unpack()`
}
\examples{
\dontrun{
  library(tidyverse)
  library(glamr)

  pano_extract(item = "mer")
}

}
