% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_extract_msd}
\alias{pano_extract_msd}
\title{Downloads Country Specific MSDs}
\usage{
pano_extract_msd(
  operatingunit = NULL,
  version = "clean",
  fiscal_year = 2021,
  quarter = 3,
  level = "psnu",
  dest_path = NULL
)
}
\arguments{
\item{operatingunit}{PEPFAR Operating Unit. Default is set to NULL for global datasets}

\item{version}{Data release version: initial or clean}

\item{fiscal_year}{Reporting Fiscal year: 4 digits year}

\item{quarter}{Reporting Quarter: Single digit quarters}

\item{level}{Org level, options are ou, psnu and sites}

\item{dest_path}{Directory path to download file. Default set to `si_path()`}
}
\description{
Downloads Country Specific MSDs
}
\examples{
\dontrun{

 pano_extract_msd(operatingunit = "Zambia",
                  version = "clean",
                  fiscal_year = 2021,
                  quarter = 3,
                  level = "site",
                  dest_path = NULL)
}

}
