% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_extract_msds}
\alias{pano_extract_msds}
\title{Downloads All Global + OU Specific MSDs}
\usage{
pano_extract_msds(
  operatingunit,
  items = "mer",
  archive = FALSE,
  dest_path = NULL,
  username = NULL,
  password = NULL,
  base_url = NULL
)
}
\arguments{
\item{operatingunit}{PEPFAR Operating Unit. Default is set to NULL for global datasets}

\item{items}{Panorama data set, default option is `mer`}

\item{archive}{Logial, should the old files be archived?}

\item{dest_path}{Directory path to download file. Default set to `si_path()`}

\item{username}{Panorama username, recommend using `pano_user()`}

\item{password}{Panorama password, recommend using `pano_pwd()`}

\item{base_url}{Panorama base url}
}
\description{
Downloads All Global + OU Specific MSDs
}
\examples{
\dontrun{
 dir_mer <- si_path()

 pano_extract_msds(operatingunit = "Zambia",
                   archive = TRUE,
                   dest_path = dir_mer)
}
}
