% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_items}
\alias{pano_items}
\title{Extract data items from url}
\usage{
pano_items(page_url, session = NULL)
}
\arguments{
\item{page_url}{Current html page url}

\item{session}{Pano active and valid session}
}
\value{
data items as data frame
}
\description{
Extract data items from url
}
\examples{
\dontrun{
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
  url <- "https://pepfar-panorama.org/forms/downloads"

  items <- pano_items(page_url = url, session = s)
}

}
