% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_session}
\alias{pano_session}
\title{Create an active session for PEPFAR Panorama}
\usage{
pano_session(username = NULL, password = NULL, base_url = NULL)
}
\arguments{
\item{username}{Username for PEPFAR Panorama Account. Recommend using `pano_user()`}

\item{password}{Password for PEPFAR Panorama Account. Recommend using `pano_pwd()`}

\item{base_url}{PEPFAR Panorama base url}
}
\value{
login session
}
\description{
Create an active session for PEPFAR Panorama
}
\examples{
\dontrun{
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
}

}
