% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_unpack}
\alias{pano_unpack}
\title{Unpack Pano files directories}
\usage{
pano_unpack(df_pano, session)
}
\arguments{
\item{df_pano}{Panorama output files items}

\item{session}{Valid and active login session}
}
\value{
unnested data frame containing output files
}
\description{
Unpack Pano files directories
}
\examples{

\dontrun{
  library(tidyverse)
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
  url <- "https://pepfar-panorama.org/forms/downloads"

  cont <- pano_content(page_url = url, session = s)

  elts <- pano_elements(page_html = cont, page_url = url)

  pano_unpack(df_pano = elts, session = s)
}

}
