% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partner_types.R
\name{partner_types}
\alias{partner_types}
\title{A code to add in partner types}
\usage{
partner_types(df)
}
\arguments{
\item{df}{You can use either a MER or financial dataset}
}
\value{
a df with the partner types provided by USAID. This includes the updates for regional partners, which in FY21 are classified as local. These adjustments can be found on the adjusted column
}
\description{
A code to add in partner types
}
