% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{pdap_bucket}
\alias{pdap_bucket}
\title{Access PDAP read/write bucket}
\usage{
pdap_bucket(type = c("read", "write"))
}
\arguments{
\item{type}{is the bucket read (default) or write?}
}
\value{
character string of AWS bucket location
}
\description{
When working with PDAP, you will need to access data from either the read or
write buckets. The read bucket ("S3_READ") is where PEPFAR Systems stores the
MSDs and the write bucket ("S3_WRITE") is where users can upload files (USAID
users will have access and write to the "usaid/" sub bucket).
}
\details{
When access from PDAP Posit Workbench, the function will access the system
environment variables `Sys.getenv("S3_READ")` or `Sys.getenv("S3_WRITE")`
where as it accessing locally, the user will need to store the read bucket
location with `set_key()`, which securely stores this information with
`keyring` (we can only write, not read from a local machine).
}
\examples{
\dontrun{
library(grabr)
s3_upload(upload_file_path,
          bucket = pdap_bucket("write"),
          prefix = "usaid/",
          access_key = pdap_access(),
          secret_key = pdap_secret())

#identify path to dataset uploaded
  path_wrkbnch <- s3_objects(bucket = pdap_bucket("write"),
                             prefix = "usaid/",
                             access_key = pdap_access(),
                             secret_key = pdap_secret()) \%>\%
    filter(str_detect(key, "Moz")) \%>\%
    pull(key)

#read
df_msd <- s3read_using(read_psd,
                       bucket = pdap_bucket("write"),
                       object = path_wrkbnch) }

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_access}()},
\code{\link{pdap_secret}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
