% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar_country_list}
\alias{pepfar_country_list}
\title{Current List of PEPFAR OUs/Countries and their ISO codes}
\format{
A data frame with 51 rows and 4 variables:
\describe{
  \item{operatingunit}{PEPFAR Operating Unit (countries + 3 regional programs)}
  \item{operatingunit_iso}{PEPFAR Operating Unit ISO-3}
  \item{countryname}{PEPFAR Country Name}
  \item{countryname_iso}{PEPFAR Country Name ISO-3}
}
}
\source{
\url{https://final.datim.org/}
}
\usage{
pepfar_country_list
}
\description{
A dataset PEPFAR Operating Units and Countries along with their ISO codes.
This is a useful dataset for having a full set of PEPFAR countries or to
align ISO codes with external datasources. Pulled from DATIM and the MSD.
}
\keyword{datasets}
