% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar_country_xwalk}
\alias{pepfar_country_xwalk}
\title{Cross-walk of Country names}
\format{
A data frame with 55 rows and 29 variables:
\describe{
  \item{iso3c}{ISO-3 Code}
  \item{continent}{Continent name}
  \item{region}{World Region Name}
  \item{region23}{Alternative name for World Region}
  \item{un_region_name}{US Region Name}
  \item{iso_name_en}{ISO Country name in English}
  \item{iso_name_fr}{SO Country name in French}
  \item{country_name_de}{Regular Expression of Country name in German}
  \item{country_name_de_regex}{Country name in German}
  \item{country_name_en}{Country name in English}
  \item{country_name_en_regex}{Regular Expression of Country name in English}
  \item{country_name_fr}{Regular Expression of Country name in French}
  \item{country_name_fr_regex}{Country name in French}
  \item{country_name_it}{Regular Expression of Country name in Italian}
  \item{country_name_it_regex}{Country name in Italian}
  \item{un_name_ar}{UN Name in Arabic}
  \item{un_name_en}{UN Name in English}
  \item{un_name_es}{UN Name in Spanish}
  \item{un_name_fr}{UN Name in French}
  \item{un_name_ru}{UN Name in ru}
  \item{un_name_zh}{UN Name in zh}
  \item{sovereignt}{rnaturalearth sovereign territory name}
  \item{admin}{rnaturalearth administrative unit name}
  \item{name}{rnaturalearth country name}
  \item{operatingunit}{PEPFAR Operating Unit (countries + 3 regional programs)}
  \item{operatingunit_iso}{PEPFAR Operating Unit ISO-3}
  \item{operatingunit_uid}{PEPFAR Operating Unit unique id from DATIM}
  \item{country}{PEPFAR Country Name}
  \item{country_uid}{PEPFAR Country unique id from DATIM}
}
}
\source{
\itemize{
  \url{https://final.datim.org/}
  \url{https://www.naturalearthdata.com/}
  \url{https://vincentarelbundock.github.io/countrycode/}
}
}
\usage{
pepfar_country_xwalk
}
\description{
A dataset PEPFAR Operating Units and Countries along with their ISO codes, alternative names from other sources.
This is a useful dataset designed to help with data cleaning / matching from different sources.
}
\keyword{datasets}
