% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_totals.R
\name{pluck_totals}
\alias{pluck_totals}
\title{Filter for Only Totals in MSD}
\usage{
pluck_totals(df)
}
\arguments{
\item{df}{MER Structured Dataset (MSD) dataframe}
}
\value{
MSD with only total numerator and denominator
}
\description{
Often times we want just want to work with totals when performing an analysis
or visualization. This function is a simple filter for Total Numerator and
Denominator in a MSD dataframe
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df_msd <- read_msd(path)
df_totals <- df_msd \%>\%
  pluck_totals() \%>\%
  clean_indicator() \%>\%
  filter(indicator \%in\% cascade_ind)
\dontshow{\}) # examplesIf}
}
