% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinf.R
\name{prinf}
\alias{prinf}
\title{Print All Rpws}
\usage{
prinf(df)
}
\arguments{
\item{df}{data frame}
}
\value{
prints out all rows rather than default 10
}
\description{
Print All Rpws
}
\examples{
\dontrun{
df_geo \%>\% prinf() }
}
