% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{print_knownissues}
\alias{print_knownissues}
\title{Print Known Issues to console}
\usage{
print_knownissues(df, store_excl = FALSE)
}
\arguments{
\item{df}{data frame from note_knownissues()}

\item{store_excl}{whether to store exclusions to Global Envir}
}
\value{
console print of known issues
}
\description{
Print Known Issues to console
}
\keyword{internal}
