% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{remove_centralsupport}
\alias{remove_centralsupport}
\title{Remove Central Support Reporting}
\usage{
remove_centralsupport(df)
}
\arguments{
\item{df}{MSD dataframe (must include indicatortype)}
}
\value{
df with central support reporting removed
}
\description{
Central Support is often desired to be removed from MER analysis. As of
FY21Q4, "CS" is a field under `indicatortype`, which can be used to
identify and exclude this sort of reporting. This function is run by
default in `resolve_knownissues`, but can be run separately if desired.
}
