% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mo.R
\name{remove_mo}
\alias{remove_mo}
\title{A Function to Remove M&O from datasets}
\usage{
remove_mo(df)
}
\arguments{
\item{df}{financial or comprehensive budget dataset}
}
\value{
a dataset excluding M&O
}
\description{
A Function to Remove M&O from datasets
}
