% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_buckets}
\alias{s3_buckets}
\title{Get S3 Buckets list}
\usage{
s3_buckets(access_key = NULL, secret_key = NULL)
}
\arguments{
\item{access_key}{S3 Access Key ID}

\item{secret_key}{S3 Secret Access Key}
}
\value{
S3 Backets list as tibble
}
\description{
Get S3 Buckets list
}
\examples{
\dontrun{
s3_buckets() }

}
