% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_excel_sheets}
\alias{s3_excel_sheets}
\title{Read sheets from S3 Objects / Excel}
\usage{
s3_excel_sheets(bucket, object_key, access_key = NULL, secret_key = NULL)
}
\arguments{
\item{bucket}{S3 Bucket}

\item{object_key}{S3 Object Key}

\item{access_key}{S3 Access Key}

\item{secret_key}{S3 Secret Key}
}
\value{
Excel sheets as data frame
}
\description{
Read sheets from S3 Objects / Excel
}
\examples{
\dontrun{
s3_objects(
    bucket = "sample-bucket",
    prefix = "ddc/xyz/ABC") \%>\%
  filter(str_detect(key, "^HFR")) \%>\%
  pull(key) \%>\%
  first() \%>\%
  s3_excel_sheets(bucket = "<sample-bucket>", object_key = .)}

}
