% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_object_type}
\alias{s3_object_type}
\title{Idenfity S3 Object type}
\usage{
s3_object_type(object_key)
}
\arguments{
\item{object_key}{S3 Object key}
}
\value{
file type: text, csv, excel, json, python, shell
}
\description{
Idenfity S3 Object type
}
\examples{
\dontrun{
s3_objects("sample-bucket") \%>\% filter(str_detect(key, "^HFR")) \%>\% s3_object_type()}

}
