% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_objects}
\alias{s3_objects}
\title{Get S3 Bucket objects list}
\usage{
s3_objects(
  bucket,
  prefix = "ddc/uat",
  n = 1000,
  unpack_keys = FALSE,
  access_key = NULL,
  secret_key = NULL,
  ...
)
}
\arguments{
\item{bucket}{S3 Bucket name}

\item{prefix}{Limit response by key}

\item{n}{Max number of record, default = 1000}

\item{unpack_keys}{Separate key colunm, default is false}

\item{access_key}{S3 Access Key ID}

\item{secret_key}{S3 Secret Access Key}

\item{...}{Additional aws.s3::get_bucket_df() options}
}
\value{
S3 Objects list as tibble
}
\description{
Get S3 Bucket objects list
}
\examples{
\dontrun{
s3_objects("sample-bucket") }

}
