% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_read_object}
\alias{s3_read_object}
\title{Read content of S3 Objects}
\usage{
s3_read_object(bucket, object_key, access_key = NULL, secret_key = NULL)
}
\arguments{
\item{bucket}{S3 Bucket name}

\item{object_key}{S3 Object key (id)}

\item{access_key}{S3 Access key id}

\item{secret_key}{S3 Secret Access key}
}
\value{
df
}
\description{
Read content of S3 Objects
}
\examples{
\dontrun{
s3_objects("sample-bucket") \%>\%
  filter(str_detect(key, "^HFR")) \%>\%
  pull(key) \%>\%
  first() \%>\%
  s3_read_object(bucket = "sample-bucket", object_key = .)}

}
