% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_remove}
\alias{s3_remove}
\title{Remove objects from S3 bucket}
\usage{
s3_remove(objects, bucket, access_key = NULL, secret_key = NULL)
}
\arguments{
\item{objects}{S3 object keys (full path)}

\item{bucket}{S3 backet name}

\item{access_key}{S3 Access Key}

\item{secret_key}{S3 Secret Key}
}
\value{
boolean
}
\description{
Remove objects from S3 bucket
}
\examples{
\dontrun{
df_objects \%>\%
  pull(key) \%>\%
  first() \%>\%
  s3_remove(objects = .,
            bucket = "test-bkt")}

}
