% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_unpack_keys}
\alias{s3_unpack_keys}
\title{Unpack Objects Key}
\usage{
s3_unpack_keys(df_objects, rmv_sysfiles = TRUE, rmv_hidden = TRUE)
}
\arguments{
\item{df_objects}{S3 objects as df}

\item{rmv_sysfiles}{Remove System object (logs, git folders, etc.)?}

\item{rmv_hidden}{Remove Hidden objects (folders/files starting with dot) ?}
}
\value{
S3 Cleaned Objects list as tibble
}
\description{
Unpack Objects Key
}
\examples{
\dontrun{
s3_objects("sample-bucket") \%>\% s3_unpack_keys()}

}
