% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_s3objects.R
\name{s3_upload}
\alias{s3_upload}
\title{Upload file to S3 Bucket}
\usage{
s3_upload(
  filepath,
  bucket,
  prefix = "",
  object = NULL,
  access_key = NULL,
  secret_key = NULL,
  ...
)
}
\arguments{
\item{filepath}{Source file path}

\item{bucket}{S3 backet name}

\item{prefix}{S3 Prefix (folder structure)}

\item{object}{Destination S3 object name (with file extention)}

\item{access_key}{S3 Access Key}

\item{secret_key}{S3 Secret Key}

\item{...}{Additional aws.S3::put_object() options}
}
\value{
boolean
}
\description{
Upload file to S3 Bucket
}
\examples{
\dontrun{
filepath \%>\% s3_upload(bucket = "test-bkt")}

}
