% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_account}
\alias{set_account}
\title{Create / Update account}
\usage{
set_account(name, keys = c("username", "password"), update = FALSE)
}
\arguments{
\item{name}{Service name of the account}

\item{keys}{List of account key names}

\item{update}{Should an existing account be overwriten}
}
\description{
Create / Update account
}
\examples{
\dontrun{
set_account(name = 's3', keys = c("access", "secret"))}

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_bucket}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
