% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_datim}
\alias{set_datim}
\title{Store DATIM credentials}
\usage{
set_datim(datim_username)
}
\arguments{
\item{datim_username}{DATIM account}
}
\value{
stores USAID email in using keyring
}
\description{
`set_datim` stores your DATIM credentials email using the `keyring` package.
This will only need to done once. After running `set_datim(user)`, you will be
promoted to enter your password through the RStudio API which will then store the
username and password in your OS credential store using `keyring`.

 The `keyring` package utilized the OS credentials store. Storing in
 a centralized, secure location allows analysts to other analysts code
 without having to manually change user names/email address to access
 DATIM or Google Drive.

 After `set_datim` has been run once, an analyst can set `load_secrets` at the
 beginning of a script, storing their email and DATIM username under Options
 for the current session.
}
\examples{
\dontrun{
set_datim("rshah") }

}
