% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_email}
\alias{set_email}
\title{Store USAID email}
\usage{
set_email(usaid_email)
}
\arguments{
\item{usaid_email}{full USAID email address}
}
\value{
stores USAID email using keyring and .Rprofile
}
\description{
`set_email` stores your USAID email using the `keyring` package. This will
only need to run once.
}
\details{
The `keyring` package utilized the OS credentials store. Storing in
 a centralized, secure location allows analysts to other analysts code
 without having to manually change user names/email address to access
 DATIM or Google Drive.

 After `set_email` has been run once, an analyst can set `load_secrets` at the
 beginning of a script, storing their email and DATIM username under Options
 for the current session.

 This function also stores the email locally in your .Rprofile, allowing to be
 used automatically as the default for `googledrive::drive_auth()` and
 `googlesheets4::gs4_auth()`
}
\examples{
\dontrun{
set_email("rshah@usaid.gov") }

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_access}()},
\code{\link{pdap_bucket}()},
\code{\link{pdap_secret}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
