% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_key}
\alias{set_key}
\title{Set value for service name}
\usage{
set_key(service, name)
}
\arguments{
\item{service}{Name of the service}

\item{name}{Name of the key}
}
\description{
Set value for service name
}
\examples{
\dontrun{
set_key(service = '<service-name>', name = '<key-name>')}

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_key}()},
\code{\link{get_keys}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{pdap_access}()},
\code{\link{pdap_bucket}()},
\code{\link{pdap_secret}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
