% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_pano}
\alias{set_pano}
\title{Store PEPFAR Panorama credentials}
\usage{
set_pano(pano_username)
}
\arguments{
\item{pano_username}{Panorama user name (email)}
}
\value{
stores Panorama username and password in using keyring
}
\description{
`set_pano` stores your PEPFAR Panoram credentials email using the `keyring` package.
This will only need to done once. After running `set_pano(user)`, you will be
promoted to enter your password through the RStudio API which will then store the
username and password in your OS credential store using `keyring`.
}
\details{
The `keyring` package utilized the OS credentials store. Storing in
 a centralized, secure location allows analysts to other analysts code
 without having to manually change user names/email address to access
 DATIM, Panorama, or Google Drive.

 After `set_pano` has been run once, an analyst can set `load_secrets` at the
 beginning of a script, storing their PEPFAR Panorama credentials under Options
 for the current session.
}
\examples{
\dontrun{
set_pano("rshah@usaid.gov") }
}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_keys}()},
\code{\link{get_key}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
