% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_paths.R
\name{set_paths}
\alias{set_paths}
\title{Set SI local folder paths}
\usage{
set_paths(
  folderpath_msd = NULL,
  folderpath_datim = NULL,
  folderpath_raster = NULL,
  folderpath_vector = NULL
)
}
\arguments{
\item{folderpath_msd}{folderpath where the MSDs are stored}

\item{folderpath_datim}{folderpath where DATIM data are store eg (org hierarchy, mech table)}

\item{folderpath_raster}{folderpath where GIS raster data are stored}

\item{folderpath_vector}{folderpath where GIS vector data are stored}
}
\value{
code chunk to paste into .Rprofile
}
\description{
Set SI local folder paths
}
\examples{
\dontrun{
set_paths <- set_path(folderpath_msd = "C:/Users/rshah/Documents/Data") }
}
