% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_s3access}
\alias{set_s3access}
\title{Store S3 Credentials - Access Key}
\usage{
set_s3access(access)
}
\arguments{
\item{access}{S3 account access key}
}
\value{
stored access key
}
\description{
`set_s3access` stores your s3 access key using the `keyring` package.
This will only need to done once. After running `set_s3access(access)`,
RStudio API which will then store the key in your OS credential store using `keyring`.
}
\examples{
\dontrun{
set_s3access("ABDCEDFFFDFDFDFD") }

}
