% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{set_s3keys}
\alias{set_s3keys}
\title{Store S3 Credentials

`set_s3keys` stores your s3 keys using the `keyring` package.
This will only need to done once. After running `set_s3keys(access, secret)`,
RStudio API which will then store the keys in your OS credential store using `keyring`.}
\usage{
set_s3keys(access, secret)
}
\arguments{
\item{access}{S3 Account Access Key}

\item{secret}{S3 Account Secret Key}
}
\value{
stored access key
}
\description{
Store S3 Credentials

`set_s3keys` stores your s3 keys using the `keyring` package.
This will only need to done once. After running `set_s3keys(access, secret)`,
RStudio API which will then store the keys in your OS credential store using `keyring`.
}
\examples{
\dontrun{
set_s3access("ABDCEDFF", "MLIZD998SD") }

}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_s3key}()},
\code{\link{load_secrets}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_s3access}()},
\code{\link{set_s3secret}()}
}
\concept{authentication}
