% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gitignore.R
\name{setup_gitignore}
\alias{setup_gitignore}
\title{Setup gitignore in Project}
\usage{
setup_gitignore()
}
\value{
adds standard gitignore, plus specific ignores
}
\description{
Setup gitignore in Project
}
\examples{
\dontrun{
  setup_gitignore() }
}
