% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_readme}
\alias{setup_readme}
\title{Setup README with Disclaimer in Project}
\usage{
setup_readme(add_disclaimer = TRUE)
}
\arguments{
\item{add_disclaimer}{should the standard disclaimer be added, default = TRUE}
}
\value{
adds/appends disclaimer to README
}
\description{
`setup_readme()` establishes a README.md with the standard USAID disclaimer
(or appends to one that currently exists). This function can be used as a
stand alone function but primarily serves `si_setup()`.
}
\examples{
\dontrun{
#standard (appends dislaimer if README exists)
  setup_readme() }
}
\seealso{
Other project setup: 
\code{\link{folder_setup}()},
\code{\link{setup_gitignore}()},
\code{\link{si_setup}()}
}
\concept{project setup}
