% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_paths.R
\name{si_path}
\alias{si_path}
\title{Access SI folderpath}
\usage{
si_path(type = "path_msd")
}
\arguments{
\item{type}{folderpath, eg "path_msd" (default), "path_datim", "path_raster", "path_vector"}
}
\value{
folderpath stored in global options
}
\description{
`si_path` accesses folder paths stored in global options to make it easier
work across analysts/machines. Analysts will first setup the paths using
`set_paths()` which then store local folder paths where larger data are stored
centrally and outside of projects.
}
\examples{
\dontrun{
#old
  list.files("C:/Users/rshah/Documents/Data", "OU_IM", full.names = TRUE)
#new
  list.files(si_path("path_msd"), "OU_IM", full.names = TRUE)
}
}
