% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{si_setup}
\alias{si_setup}
\title{Default SI Project Setup}
\usage{
si_setup()
}
\value{
creates folders, readme, and gitignore
}
\description{
`si_setup()` combines three function - `folder_setup()`,
`setup_gitignore()`, and `setup_readme()` to create the base OHA/SI project
structure with the necessary folders, disclaimers, and ignored files.
}
\seealso{
Other project setup: 
\code{\link{folder_setup}()},
\code{\link{setup_gitignore}()},
\code{\link{setup_readme}()}
}
\concept{project setup}
