% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_info.R
\name{source_info}
\alias{source_info}
\title{Extract MSD Source Information}
\usage{
source_info(path, type, return = "source")
}
\arguments{
\item{path}{path to the folder containing MSDs or specific MSD file}

\item{type}{not required unless providing a folder in `path`; default = "OU_IM_FY19"
other examples include: "PSNU_IM", "NAT_SUBNAT", "PSNU", "Financial"}

\item{return}{from the info, what should be returned; default = "source"
other options are: "period", "fiscal_year", "quarter"}
}
\value{
vector of information related to what is being asked in `return`
}
\description{
This function is used primarily to extract the data from the source file of
a MER Structured Dataset, MER NAT_SUBNAT StructuredDataset, Financial
Structure Dataset, or DATIM Genie export. It can also be used to extact
information from the filename about the fiscal year, quarter, or period.
}
\examples{
\dontrun{
 source_info() #works if you have stored path to the MSD folder via glamr::set_paths()
 source_info("../Data", type = "PSNUxIM")
 source_info("../Data", type = "PSNUxIM", return = "period")
 source_info("../Downloads/Genie_PSNU_IM_Jupiter_Daily_c9f5889f-86c9-44e7-ab63-fa86c587d251.zip")
 source_info("../Data/MER_Structured_Datasets_NAT_SUBNAT_FY15-21_20210618_v2_1.rds") }

\dontrun{
library(tidyverse)
library(glamr)
library(glue)

df <- si_path() \%>\%
  return_latest("OU_IM_FY19") \%>\%
  read_rds()

df_viz <- df \%>\%
  filter(operatingunit == "Saturn",
         indicator == "TX_NEW",
         standardizeddisaggregate == "Total Numerator") \%>\%
  count(fiscal_year, wt = targets, name = "targets")

df_viz \%>\%
  ggplot(aes(fiscal_year, targets)) +
  geom_col() +
  labs(caption = glue("Source: {source_info()}")) }
}
