% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{squish_knownissues}
\alias{squish_knownissues}
\title{Squish Known Issues}
\usage{
squish_knownissues(df)
}
\arguments{
\item{df}{df output from flag_knownissues()}
}
\value{
dataframe with known issues data removed
}
\description{
Squish Known Issues
}
\keyword{internal}
