% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{temp_folder}
\alias{temp_folder}
\title{Generate Temporary Folder}
\usage{
temp_folder(launch = FALSE, quiet = FALSE)
}
\arguments{
\item{launch}{do you want to launch the temp folder in the Windows Explorer?
default = FALSE}

\item{quiet}{suppresses the output message related to the folder creation
and location, for use in sub functions, default = FALSE)}
}
\value{
creates a temp directory and stores it as `folderpath_tmp`
}
\description{
`temp_folder` created a temporary folder in your AppData directory, which
will be automatically removed after you close your RStudio session.
}
\examples{
\dontrun{
load_secrets()
temp_folder(launch = TRUE)
purrr::walk2(.x = df_googlefiles$id,
             .y = df_googlefiles$filename,
             .f = ~googledrive::drive_download(googledrive::as_id(.x),
                                               file.path(folderpath_tmp, .y))) }
}
