% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_paths.R
\name{validate_path}
\alias{validate_path}
\title{Validation Check to ensure valid folder path}
\usage{
validate_path(path, type)
}
\arguments{
\item{path}{folder path}

\item{type}{path type}
}
\description{
Validation Check to ensure valid folder path
}
\keyword{internal}
