% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_knownissues.R
\name{validate_structure}
\alias{validate_structure}
\title{Validate supplied dataframe to ensure valid structure}
\usage{
validate_structure(df)
}
\arguments{
\item{df}{data from get_knownissues()}
}
\value{

}
\description{
Validate supplied dataframe to ensure valid structure
}
\keyword{internal}
