% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_column}
\alias{clean_column}
\title{Clean column data}
\usage{
clean_column(.data, colname = "psnu")
}
\arguments{
\item{.data}{MSD Datasets}

\item{colname}{Name of the column(s)}
}
\value{
Cleaned MSD DataFrame
}
\description{
Clean column data
}
\examples{
\dontrun{
 df_msd \%>\% clean_column(colname = "psnu") }
}
