% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_countries}
\alias{clean_countries}
\title{Clean OU/Country names to match PEPFAR Data}
\usage{
clean_countries(.data, colname = "admin")
}
\arguments{
\item{.data}{MSD Datasets}

\item{colname}{Column name to be updated}
}
\value{
Cleaned DataFrame
}
\description{
Clean OU/Country names to match PEPFAR Data
}
