% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_countries}
\alias{clean_countries}
\title{Clean OU/Country names to match PEPFAR Data

`clean_countries` is used to adjust Natural Earth country names to match
PEPFAR ones.}
\usage{
clean_countries(.data, colname = "admin", language = "en")
}
\arguments{
\item{.data}{Reference Datasets}

\item{colname}{Column name to be updated}

\item{language}{language of reference, default is set to `en`. Options are: `fr`, `de`, `es`, `ar`}
}
\value{
Cleaned DataFrame
}
\description{
Clean OU/Country names to match PEPFAR Data

`clean_countries` is used to adjust Natural Earth country names to match
PEPFAR ones.
}
\examples{
\dontrun{
 library(sf)
 library(rnaturalearth)
 library(glamr)

 spdf <- ne_countries(type = "sovereignty", scale = 110, returnclass = "sf") \%>\%
   sf::st_drop_geometry() \%>\%
   dplyr::select(sovereignt, admin, name, adm0_a3) \%>\%
   glamr::clean_countries(colname = "admin") }

}
\concept{Column clean up}
