% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_indicator}
\alias{clean_indicator}
\title{Clean indicators}
\usage{
clean_indicator(df)
}
\arguments{
\item{df}{MSD data frame}
}
\value{
indicators with denominator have _D suffix
}
\description{
Clean indicators
}
\examples{
\dontrun{
df <- df \%>\%
filter(indicator == "TX_PVLS",
       standardizeddisaggregate \%in\% c("Total Numerator", "Total Denominator")) \%>\%
clean_indicator() }
}
