% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_columns.R
\name{clean_psnu}
\alias{clean_psnu}
\title{Clean PSNU column data}
\usage{
clean_psnu(.data)
}
\arguments{
\item{.data}{MSD Datasets}
}
\value{
Cleaned MSD DataFrame
}
\description{
Clean PSNU column data
}
\examples{
\dontrun{
 df_msd \%>\% clean_psnu() }
}
\seealso{
Other column munging: 
\code{\link{clean_agency}()},
\code{\link{clean_column}()},
\code{\link{clean_indicator}()}
}
\concept{column munging}
