% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_date_period.R
\name{convert_datim_pd_to_qtr}
\alias{convert_datim_pd_to_qtr}
\title{Convert DATIM CY Quarter/Period to FY Quarter}
\usage{
convert_datim_pd_to_qtr(df)
}
\arguments{
\item{df}{dataframe from \code{extract_datim()}}
}
\value{
Convert periods from long CY dates to PEPFAR standard FY
}
\description{
Convert a period a DATIM API in the format of FY22Q1 or FY22 (for
targets/cumulative). This function is built into `extract_datim`.
}
\examples{
\dontrun{
df <- tibble::tibble(Periods = c("Jan to Mar 2019", "Oct 2018 to Sep 2019"))
df <- convert_datim_pd_to_qtr(df) }
}
\seealso{
[set_datim()] to store DATIM authentication;
[load_secrets()] to store DATIM authentication;
[extract_datim()] to run API (which incorporates convert_datim_pd_to_qtr)

Other period: 
\code{\link{convert_date_to_qtr}()},
\code{\link{convert_fy_qtr_to_pd}()},
\code{\link{convert_qtr_to_date}()}
}
\concept{period}
