% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datim.R
\name{datim_dimension}
\alias{datim_dimension}
\title{Get PEPFAR/DATIM Dimension ID}
\usage{
datim_dimension(
  name,
  url = "https://final.datim.org/api/dimensions",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{name}{Dimension name}

\item{url}{DATIM API End point, default value is `https://final.datim.org/api/dimensions`}

\item{username}{DATIM Account Username, recommended using `datim_user()`}

\item{password}{DATIM Account passward, recommended using `datim_pwd()`}
}
\value{
dimension uid
}
\description{
Get PEPFAR/DATIM Dimension ID
}
\examples{
\dontrun{
  library(glamr)
  datim_dimension("OU Level")
}

}
