% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{datim_pwd}
\alias{datim_pwd}
\title{Return DATIM password}
\usage{
datim_pwd()
}
\value{
access DATIM password from keyring
}
\description{
To setup/store, run `glamr::set_datim()`.
}
\examples{
\dontrun{
load_secrets()
ou_table <- datim_outable(datim_user(), datim_pwd()) }
}
\seealso{
Other authentication: 
\code{\link{datim_user}()},
\code{\link{get_s3key}()},
\code{\link{load_secrets}()},
\code{\link{pano_pwd}()},
\code{\link{pano_user}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
