% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{datim_user}
\alias{datim_user}
\title{Return DATIM username}
\usage{
datim_user()
}
\value{
access DATIM username from keyring
}
\description{
Return DATIM username
}
\examples{
\dontrun{
load_secrets()
ou_table <- datim_outable(datim_user(), datim_pwd()) }
}
