% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_drivefile.R
\name{export_drivefile}
\alias{export_drivefile}
\title{Update files}
\usage{
export_drivefile(
  filename,
  to_drive,
  to_folder = NULL,
  add_folder = TRUE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
Googledrive file(s) id(s)
}
\description{
Update files
}
\examples{
\dontrun{
 library(glamr)

 list.files("./Graphics", "NIGERIA", full.names = TRUE) \%>\%
      export_drivefile(filename = .,
                       to_drive = "<path-id>",
                       to_folder = "FY99Q4/VL Suppression",
                       add_folder = TRUE)

}

}
