% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_drivefile.R
\name{export_drivefile}
\alias{export_drivefile}
\title{Update files}
\usage{
export_drivefile(
  filename,
  to_drive,
  to_folder = NULL,
  add_folder = TRUE,
  file_type = "png",
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{filename}{Full name of the file to be uploaded}

\item{to_drive}{Google drive id}

\item{to_folder}{Google drive sub-folder}

\item{add_folder}{If TRUE, add sub-folders if they are not present}

\item{file_type}{File type (extension)}

\item{overwrite}{If yes, existing files will be overwritten}

\item{...}{Additional parameters to be passed to `googledrive::drive_upload()`}
}
\value{
Googledrive file(s) id(s)
}
\description{
Update files
}
\examples{
\dontrun{
 library(glamr)

 list.files("./Graphics", "NIGERIA", full.names = TRUE) \%>\%
      export_drivefile(filename = .,
                       to_drive = "<path-id>",
                       to_folder = "FY99Q4/VL Suppression",
                       add_folder = TRUE)

}

}
