% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{folder_setup}
\alias{folder_setup}
\title{Setup Folder Structure in Project}
\usage{
folder_setup(
  folder_list = list("Data", "Images", "Scripts", "Dataout", "GIS", "Documents",
    "Graphics", "markdown")
)
}
\arguments{
\item{folder_list}{list of folders to install}
}
\description{
Setup Folder Structure in Project
}
\examples{
\dontrun{
#standard
  folder_setup()
#specific
  fldrs <- c("Data", "Tableau", "AI")
  folder_setup(fldrs)
}
}
