% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_ouorglabel}
\alias{get_ouorglabel}
\title{Identify OU/Org Label}
\usage{
get_ouorglabel(
  operatingunit,
  country = NULL,
  org_level = 4,
  username = NULL,
  password = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{operatingunit}{Operating unit}

\item{country}{Country name}

\item{org_level}{OU Org level, default is set to 4, PSNU}

\item{username}{Datim account username}

\item{password}{Datim account password}

\item{baseurl}{Datim base url}
}
\value{
Org level label
}
\description{
Identify OU/Org Label
}
\examples{
\dontrun{
  library(glamr)

  get_ouorglabel(operatingunit = "Zambia", org_level = 5)
}

}
