% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{identify_levels}
\alias{identify_levels}
\title{Identify Facility/Community levels in org hierarchy}
\usage{
identify_levels(username, password, baseurl = "https://final.datim.org/")
}
\arguments{
\item{username}{DATIM Username, defaults to using `datim_user()` if blank}

\item{password}{DATIM password, defaults to using `datim_pwd()` if blank}

\item{baseurl}{base url for the API, default = https://final.datim.org/}
}
\description{
`identify_levels` pulls from DATIM to return a dataframe with all PEPFAR
Operating Units and countries with their ISO codes and hierarhcy levels.
This is one of two components that feeds into `get_outable`.
}
\details{
To access the UIDs, the user will need to have a DATIM account. You can
take advantage of storing you credentials locally in a secure way
using `set_datim`.
}
\examples{
\dontrun{
 #table for all OUs
  load_secrets()
  identify_levels() }
}
\seealso{
[set_datim()] to store DATIM authentication;
  [load_secrets()] to load credentials into session
}
