% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_drivefile.R
\name{import_drivefile}
\alias{import_drivefile}
\title{Import file from a Google Drive folder}
\usage{
import_drivefile(drive_folder, filename, folderpath = "Data", zip = TRUE)
}
\arguments{
\item{drive_folder}{Google id for Google Drive Folder}

\item{filename}{exact name of file on Googl Drive to download}

\item{folderpath}{path where you want file stored, default = "Data"}

\item{zip}{should the file be zipped? default = TRUE}
}
\value{
stores file from Google Drive as a zipped file
}
\description{
`import_drivefile` is a wrapper around `googledrive::drive_download`, useful
for pulling multiple files from a given Google Drive folder (with a Google
ID provided) to download by default to the Data folder of a project.
}
\examples{
\dontrun{
library(googledrive)
googledrive::drive_auth()
fldr <- "Spp-y8DYsdRTrzDqUmK4fX5v"
import_drivefile(fldr, "TestFile.csv")
}
}
