% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_loaded}
\alias{is_loaded}
\title{Test if service is loaded in current session}
\usage{
is_loaded(service = c("datim", "email"))
}
\arguments{
\item{service}{account, either "datim" or "email"}
}
\value{
A boolean
}
\description{
Test if service is loaded in current session
}
