% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_stored}
\alias{is_stored}
\title{Test if service is stored in credential manager}
\usage{
is_stored(service = c("datim", "email"))
}
\arguments{
\item{service}{account, either "datim" or "email"}
}
\value{
A boolean
}
\description{
Test if service is stored in credential manager
}
