% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{load_secrets}
\alias{load_secrets}
\title{Load credentials}
\usage{
load_secrets()
}
\value{
stores Google and DATIM credentials in session
}
\description{
#' @description
 `load_secrets` should be set at the beginning of a script to store your
 email and DATIM user name under Options for the current session. This allows
 analysts to more easily share their scripts without having to manually update
 or remove use names.
}
\details{
To initiall store your Google email and DATIM credentials,
 you will first need to
 run `set_email()`, `set_datim()`, `set_s3access()` and `set_s3secret()`

 `load_secrets` utilizes `keyring` package to access the OS credentials store.
 Storing in a centralized, secure location allows analysts to other analysts code
 without having to manually change user names/email address to access
 DATIM or Google Drive.
}
\examples{
\dontrun{
load_secrets()
user <- getOption("datim")
ou_table <- datim_outable(user, datim_pwd()) }
}
