% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pano.R
\name{pano_download}
\alias{pano_download}
\title{Download file from PEPFAR Panorama}
\usage{
pano_download(item_url, session, dest = NULL, uncompress = FALSE)
}
\arguments{
\item{item_url}{URL for the item to be downlaoded}

\item{session}{Login session}

\item{dest}{Location and name of the destination file}

\item{uncompress}{If yes, the downloaded zip file will be decompressed. Default is FALSE}
}
\value{
file content as binary
}
\description{
Download file from PEPFAR Panorama
}
\examples{
\dontrun{
  library(tidyverse)
  library(glamr)

  s <- pano_session("<my-pano-user>", "<my-password>")
  url <- "https://pepfar-panorama.org/forms/downloads"

  cont <- pano_content(page_url = url, session = s)

  elts <- pano_elements(page_html = cont, page_url = url)

  f_url <- elts \%>\% filter(type == "file zipfile") \%>\% pull(path) \%>\% first()

  pano_download(item_url = f_url, session = s, dest = "./Data/")}

}
