% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secrets.R
\name{pano_pwd}
\alias{pano_pwd}
\title{Return PEPFAR Panorama password}
\usage{
pano_pwd()
}
\value{
access Panorama password from keyring
}
\description{
To setup/store, run `glamr::set_pano()`.
}
\seealso{
Other authentication: 
\code{\link{datim_pwd}()},
\code{\link{datim_user}()},
\code{\link{get_account}()},
\code{\link{get_keys}()},
\code{\link{get_key}()},
\code{\link{get_s3key}()},
\code{\link{get_services}()},
\code{\link{load_secrets}()},
\code{\link{pano_user}()},
\code{\link{set_account}()},
\code{\link{set_datim}()},
\code{\link{set_email}()},
\code{\link{set_key}()},
\code{\link{set_pano}()},
\code{\link{set_s3keys}()}
}
\concept{authentication}
