% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar_country_list}
\alias{pepfar_country_list}
\title{Current List of PEPFAR OUs/Countries and their ISO codes}
\format{
A data frame with 55 rows and 7 variables:
\describe{
  \item{operatingunit}{PEPFAR Operating Unit (countries + 3 regional programs)}
  \item{operatingunit_iso}{PEPFAR Operating Unit ISO-3}
  \item{operatingunit_uid}{PEPFAR Operating Unit unique id from DATIM}
  \item{country}{PEPFAR Country Name}
  \item{country_iso}{PEPFAR Country Name ISO-3}
  \item{country_uid}{PEPFAR Country unique id from DATIM}
  \item{pepfar_accel}{Is the country prioritized in the 8+1+1 group defined by SGAC?}
}
}
\source{
\url{https://final.datim.org/}
}
\usage{
pepfar_country_list
}
\description{
A dataset PEPFAR Operating Units and Countries along with their ISO codes.
This is a useful dataset for having a full set of PEPFAR countries or to
align ISO codes with external datasources. Pulled from DATIM and the FSD.
}
\details{
The list of PEPFAR acceleration countries was defined by Amb. Nkengasong
during a DP's retreat for the [Zaidi 2023-06-08 re: Moving
countries to green!]. These are countries "where enhanced attention and
focus might help 'move the dial' on achieving and sustaining the
UNAIDS 95-95-95 targets by 2025"
}
\keyword{datasets}
