% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar_data_calendar}
\alias{pepfar_data_calendar}
\title{Current PEPFAR Data Calendar}
\format{
A data frame with 16 rows and 6 variables:
\describe{
  \item{fiscal_year}{fiscal year, start = October}
  \item{quarter}{fiscal quarter, integer 1-4}
  \item{type}{data release type, initial or clean}
  \item{entry_open}{date entry begins into DATIM}
  \item{entry_close}{date DATIM is closed and data are frozen}
  \item{msd_release}{date the MSD is released on PEPFAR Panorma}
}
}
\source{
\url{https://datim.zendesk.com/hc/en-us/articles/115001940503-PEPFAR-Data-Calendar}
}
\usage{
pepfar_data_calendar
}
\description{
A dataset the dates for the release of the MER Structured Dataset
}
\keyword{datasets}
