% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinf.R
\name{prinf}
\alias{prinf}
\title{Print All Rows}
\usage{
prinf(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{Any other valid option for `print()`.}
}
\value{
prints out all rows rather than default 10 rows.
}
\description{
`prinf` is a wrapper around `print` that returns all rows rather than
just the first 10 by default.
}
\examples{
\dontrun{
df_geo \%>\% prinf() }
}
