% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mo.R
\name{remove_mo}
\alias{remove_mo}
\title{Remove M&O funding}
\usage{
remove_mo(df)
}
\arguments{
\item{df}{financial or comprehensive budget dataset}
}
\value{
a dataset excluding M&O
}
\description{
When working with financial data, its often useful to remove M&O
funding from the data.
}
