% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sch.R
\name{remove_sch}
\alias{remove_sch}
\title{A Function to Pull Out Supply Chain Data}
\usage{
remove_sch(df, poc = c("SCH", "SGAC"))
}
\arguments{
\item{df}{this can be either a financial strucutured dataset or an MSD}

\item{poc}{you can choose to filter either the SCH list, the SGAC list; default is both}
}
\value{
a df without supply chain mechanisms
}
\description{
A Function to Pull Out Supply Chain Data
}
