% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_latest.R
\name{return_latest}
\alias{return_latest}
\title{Return Latest File}
\usage{
return_latest(folderpath, pattern, quiet = FALSE, ...)
}
\arguments{
\item{folderpath}{path to folder where file(s) are located.}

\item{pattern}{pattern in file name, regex expressions. If not parttern is
provided, the last file in the folder will be returned.}

\item{quiet}{suppresses the output message related to the file name creation,
for use in sub functions, default = FALSE}

\item{...}{Any other valid option for `base::list.files()`.}
}
\value{
a vector of the full filepath for the most recent version of a file stub
}
\description{
`return_latest` checks for a pattern in a folder and provides the most
recent file bases on the time the file was modified
}
\examples{
\dontrun{
file_stub <- "MER_Structured_Datasets_OU_IM_FY18-20"
filepath <- return_latest("Data", file_stub)
df <- read_rds(filepath) }
}
