% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_latest.R
\name{return_latest}
\alias{return_latest}
\title{Return Latest}
\usage{
return_latest(folderpath, pattern)
}
\arguments{
\item{folderpath}{path to folder where file(s) are located}

\item{pattern}{pattern in file name, regex expressions}
}
\value{
the fill filepath for the most recent version of a file stub
}
\description{
Return Latest
}
\examples{
\dontrun{
file_stub <- "MER_Structured_Datasets_OU_IM_FY18-20"
filepath <- return_latest("Data", file_stub)
df <- read_rds(filepath) }
}
