% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_gitignore}
\alias{setup_gitignore}
\title{Setup gitignore in Project}
\usage{
setup_gitignore()
}
\value{
adds standard gitignore, plus specific ignores
}
\description{
`setup_gitignore()` creates a .gitignore file (or appends to an existing)
one) the standard file types/folders that should not be published to GitHub
due to the sensitive nature of PEPFAR data. This function can be used as a
stand alone function but primarily serves `si_setup()`.
}
\examples{
\dontrun{
  setup_gitignore() }
}
\seealso{
Other project setup: 
\code{\link{setup_readme}()},
\code{\link{si_setup}()}
}
\concept{project setup}
