% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_readme.R
\name{setup_readme}
\alias{setup_readme}
\title{Setup README with Disclaimer in Project}
\usage{
setup_readme(add_disclaimer = TRUE)
}
\arguments{
\item{add_disclaimer}{should the standard disclaimer be added, default = TRUE}
}
\value{
adds/appends disclaimer to README
}
\description{
Setup README with Disclaimer in Project
}
\examples{
\dontrun{
#standard (appends dislaimer if README exists)
  setup_readme() }
}
